Restructure files, tests passing
