+++ xlron/environments/gn_model/gn_model_test.py
@@ -104,5 +104,6 @@ class RSAGNModelTest(parameterized.TestCase):
         chex.assert_trees_all_close(env_state.link_snr_array, expected)
 
 if __name__ == '__main__':
+    os.environ['XLA_FLAGS'] = "--xla_force_host_platform_device_count=4"
     jax.config.update('jax_numpy_rank_promotion', 'raise')
     absltest.main()
