+++ xlron/environments/vone/vone.py
@@ -9,7 +9,7 @@ import jraph
 from gymnax.environments import environment, spaces
 from xlron.environments.env_funcs import (
     init_vone_request_array, init_link_slot_array, init_link_slot_mask, init_link_slot_departure_array,
-    check_vone_action, undo_action_rsa, finalise_vone_action, generate_vone_request, mask_slots,
+    check_vone_action, undo_action_rsa, finalise_vone_action, generate_vone_request, mask_slots, init_traffic_matrix,
     init_node_capacity_array, init_node_mask, init_node_resource_array, init_node_departure_array, init_values_nodes,
     init_action_counter, init_action_history, update_action_history, decrease_last_element, undo_node_action,
     init_virtual_topology_patterns, mask_nodes, init_graph_tuple, format_vone_slot_request, implement_vone_action,
@@ -22,7 +22,13 @@ class VONEEnv(environment.Environment):
     """This environment simulates the Virtual Optical Network Embedding (VONE) problem.
 
     """
-    def __init__(self, params: VONEEnvParams, virtual_topologies=["3_ring"]):
+    def __init__(self,
+                 key: chex.PRNGKey,
+                 params: VONEEnvParams,
+                 virtual_topologies=["3_ring"],
+                 traffic_matrix: chex.Array = None,
+                 list_of_requests: chex.Array = None,
+                 laplacian_matrix: chex.Array = None,):
         super().__init__()
         state = VONEEnvState(
             current_time=0,
@@ -46,8 +52,11 @@ class VONEEnv(environment.Environment):
             full_link_slot_mask=init_link_slot_mask(params),
             accepted_services=0,
             accepted_bitrate=0.,
+            total_bitrate=0.,
+            list_of_requests=list_of_requests,
+            traffic_matrix=traffic_matrix if traffic_matrix is not None else init_traffic_matrix(key, params),
         )
-        self.initial_state = state.replace(graph=init_graph_tuple(state, params))
+        self.initial_state = state.replace(graph=init_graph_tuple(state, params, laplacian_matrix, exclude_source_dest=True))
 
     @partial(jax.jit, static_argnums=(0, 4))
     def step(
@@ -111,7 +120,7 @@ class VONEEnv(environment.Environment):
         check = check_vone_action(state, remaining_actions, total_requested_nodes)
         state, reward = jax.lax.cond(
             check,  # Fail if true
-            lambda x: (undo_action_rsa(undo_node_action(x)), self.get_reward_failure(x)),
+            lambda x: (undo_action_rsa(undo_node_action(x), params), self.get_reward_failure(x)),
             lambda x: jax.lax.cond(
                 jnp.any(remaining_actions <= 1),  # Final action
                 lambda xx: (finalise_vone_action(xx), self.get_reward_success(xx)),  # Finalise actions if complete
@@ -137,6 +146,7 @@ class VONEEnv(environment.Environment):
         else:
             done = self.is_terminal(state, params)
         # Update graph tuple
+        # TODO - get the update graph tuple working with VONE
         state = state.replace(graph=init_graph_tuple(state, params))
         info = {}
         return self.get_obs(state), state, reward, done, info
