+++ xlron/environments/vone/vone_test.py
@@ -33,46 +33,47 @@ def keys_test_setup():
 def settings_vone_4node():
     return dict(load=100, k=2, topology_name="4node", link_resources=4, max_requests=10, mean_service_holding_time=10,
                 node_resources=4, virtual_topologies=["3_ring"], min_node_resources=1, max_node_resources=1,
-                values_bw=[0], slot_size=1)
+                values_bw=[1], slot_size=1, env_type="vone", guardband=0)
 
 
 def settings_vone_4node_mod():
     return dict(load=100, k=2, topology_name="4node", link_resources=4, max_requests=10, mean_service_holding_time=10,
-                node_resources=4, virtual_topologies=["3_ring"], min_node_resources=1, max_node_resources=1,
-                env_type="rmsa")
+                node_resources=4, virtual_topologies=["3_ring"], min_node_resources=1, max_node_resources=1, consider_modulation_format=True,
+                env_type="vone", guardband=0)
 
 
 def vone_4node_test_setup():
     key = jax.random.PRNGKey(0)
-    env, params = make(settings_vone_4node())
+    env, params = make(settings_vone_4node(), log_wrapper=False)
     obs, state = env.reset(key, params)
     return key, env, obs, state, params
 
 
 def vone_4node_mod_test_setup():
     key = jax.random.PRNGKey(0)
-    env, params = make(settings_vone_4node_mod())
+    env, params = make(settings_vone_4node_mod(), log_wrapper=False)
     obs, state = env.reset(key, params)
     return key, env, obs, state, params
 
 
 def vone_nsfnet_16_test_setup():
     key = jax.random.PRNGKey(0)
-    settings_vone_nsfnet_16 = dict(load=100, k=5, topology_name="nsfnet", link_resources=16, max_requests=10,
+    settings_vone_nsfnet_16 = dict(load=100, k=5, topology_name="nsfnet_deeprmsa_undirected", link_resources=16, max_requests=10,
                                    values_bw=[1, 2, 3], slot_size=1,  consider_modulation_format=False,
                                    mean_service_holding_time=10, node_resources=4, virtual_topologies=["3_ring"],
-                                   min_node_resources=1, max_node_resources=2)
-    env, params = make(settings_vone_nsfnet_16)
+                                   min_node_resources=1, max_node_resources=2, env_type="vone")
+    env, params = make(settings_vone_nsfnet_16, log_wrapper=False)
     obs, state = env.reset(key, params)
     return key, env, obs, state, params
 
 
 def vone_nsfnet_16_mod_test_setup():
     key = jax.random.PRNGKey(0)
-    settings_vone_nsfnet_16_mod = dict(load=100, k=5, topology_name="nsfnet", link_resources=16, max_requests=10,
+    settings_vone_nsfnet_16_mod = dict(load=100, k=5, topology_name="nsfnet_deeprmsa_undirected", link_resources=16, max_requests=10,
                                    mean_service_holding_time=10, node_resources=4, virtual_topologies=["3_ring"],
-                                   min_node_resources=1, max_node_resources=2, consider_modulation_format=True)
-    env, params = make(settings_vone_nsfnet_16_mod)
+                                   min_node_resources=1, max_node_resources=2, consider_modulation_format=True,
+                                       env_type="vone")
+    env, params = make(settings_vone_nsfnet_16_mod, log_wrapper=False)
     obs, state = env.reset(key, params)
     return key, env, obs, state, params
 
@@ -146,6 +147,131 @@ class RemoveExpiredNodeRequestsTest(parameterized.TestCase):
         state = state.replace(current_time=10e4)
         updated_state = self.variant(remove_expired_node_requests)(state)
         chex.assert_trees_all_close(updated_state.node_departure_array, expected)
+
+
+class UpdateNodeDepartureTest(parameterized.TestCase):
+
+    @chex.all_variants()
+    @parameterized.named_parameters(
+        ('case_base',
+         jnp.array([jnp.inf, jnp.inf, jnp.inf, jnp.inf, jnp.inf]), 3, 0.1,
+         jnp.array([jnp.inf, jnp.inf, jnp.inf, 0.1, jnp.inf])),
+    )
+    def test_update_node_departure(self, node_row, inf_index, value, expected):
+        updated_node_row = self.variant(update_node_departure)(node_row, inf_index, value)
+        chex.assert_trees_all_close(updated_node_row, expected)
+
+
+class UpdateSelectedNodeDepartureTest(parameterized.TestCase):
+
+    @chex.all_variants()
+    @parameterized.named_parameters(
+        ('case_base',
+         jnp.array([jnp.inf, jnp.inf, jnp.inf, jnp.inf, jnp.inf]), 1, 3, 0.1,
+         jnp.array([jnp.inf, jnp.inf, jnp.inf, 0.1, jnp.inf])),
+        ('case_not_selected',
+          jnp.array([jnp.inf, jnp.inf, jnp.inf, jnp.inf, jnp.inf]), 0, 3, 0.1,
+          jnp.array([jnp.inf, jnp.inf, jnp.inf, jnp.inf, jnp.inf])),
+    )
+    def test_update_selected_node_departure(self, node_row, node_selected, inf_index, value, expected):
+        updated_node_row = self.variant(update_selected_node_departure)(node_row, node_selected, inf_index, value)
+        chex.assert_trees_all_close(updated_node_row, expected)
+
+
+class VmapUpdateNodeDepartureTest(parameterized.TestCase):
+
+    @chex.all_variants()
+    @parameterized.named_parameters(
+        ('case_base',
+         jnp.array([[jnp.inf, jnp.inf, jnp.inf, jnp.inf, jnp.inf],
+                   [jnp.inf, jnp.inf, jnp.inf, jnp.inf, jnp.inf],
+                   [jnp.inf, jnp.inf, jnp.inf, jnp.inf, jnp.inf],
+                   [jnp.inf, jnp.inf, jnp.inf, jnp.inf, jnp.inf]]), jnp.array([1,0,0,0]), 0.1,
+         jnp.array([[0.1, jnp.inf, jnp.inf, jnp.inf, jnp.inf],
+                    [jnp.inf, jnp.inf, jnp.inf, jnp.inf, jnp.inf],
+                    [jnp.inf, jnp.inf, jnp.inf, jnp.inf, jnp.inf],
+                    [jnp.inf, jnp.inf, jnp.inf, jnp.inf, jnp.inf]])),
+        ('case_occupied',
+         jnp.array([[jnp.inf, 1, 1, jnp.inf, jnp.inf],
+                    [1, jnp.inf, 1, jnp.inf, jnp.inf],
+                    [jnp.inf, jnp.inf, jnp.inf, jnp.inf, jnp.inf],
+                    [1, 1, 1, jnp.inf, jnp.inf]]), jnp.array([1, 1, 1, 1]), 0.99,
+         jnp.array([[0.99, 1, 1, jnp.inf, jnp.inf],
+                    [1, 0.99, 1, jnp.inf, jnp.inf],
+                    [0.99, jnp.inf, jnp.inf, jnp.inf, jnp.inf],
+                    [1, 1, 1, 0.99, jnp.inf]])),
+    )
+    def test_vmap_update_node_departure(self, node_departure_array, selected_nodes, value, expected):
+        updated_link_array = self.variant(vmap_update_node_departure)(node_departure_array, selected_nodes, value)
+        chex.assert_trees_all_close(updated_link_array, expected)
+
+
+class UpdateNodeArrayTest(parameterized.TestCase):
+    @chex.all_variants()
+    @parameterized.named_parameters(
+        ('case_base', jnp.arange(10), jnp.full(10, 10), 0, 4, jnp.array([6, 10, 10, 10, 10, 10, 10, 10, 10, 10])),
+    )
+    def test_update_node_array(self, node_indices, node_array, node, request, expected):
+        updated_node_array = self.variant(update_node_array)(node_indices, node_array, node, request)
+        chex.assert_trees_all_close(updated_node_array, expected)
+
+
+class UpdateNodeResourcesTest(parameterized.TestCase):
+
+    @chex.all_variants()
+    @parameterized.named_parameters(
+        ('case_base', jnp.zeros(10), 4, 6, jnp.array([0, 0, 0, 0, 6, 0, 0, 0, 0, 0])),
+    )
+    def test_update_node_resources(self, node_row, zero_index, value, expected):
+        updated_node_resources = self.variant(update_node_resources)(node_row, zero_index, value)
+        chex.assert_trees_all_close(updated_node_resources, expected)
+
+
+class UpdateSelectedNodeResourcesTest(parameterized.TestCase):
+
+    @chex.all_variants()
+    @parameterized.named_parameters(
+        ('case_base', jnp.zeros(10), 6, 4, jnp.array([0, 0, 0, 0, 6, 0, 0, 0, 0, 0])),
+    )
+    def test_update_selected_node_resources(self, node_row, request, zero_index, expected):
+        updated_node_resources = self.variant(update_selected_node_resources)(node_row, request, zero_index)
+        chex.assert_trees_all_close(updated_node_resources, expected)
+
+
+class VmapUpdateNodeResourcesTest(parameterized.TestCase):
+
+    @chex.all_variants()
+    @parameterized.named_parameters(
+        ('case_base',
+         jnp.full((4, 10), 0), jnp.array([4,5,6,7]),
+         jnp.array([[4, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+                   [5, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+                   [6, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+                   [7, 0, 0, 0, 0, 0, 0, 0, 0, 0]])),
+        ('case_occupied',
+         jnp.array([[4, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+                    [5, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+                    [6, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+                    [7, 0, 0, 0, 0, 0, 0, 0, 0, 0]]),
+         jnp.array([4, 0, 0, 7]),
+         jnp.array([[4, 4, 0, 0, 0, 0, 0, 0, 0, 0],
+                    [5, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+                    [6, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+                    [7, 7, 0, 0, 0, 0, 0, 0, 0, 0]])),
+        ('case_negatives',
+         jnp.array([[-1, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+                    [-1, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+                    [-1, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+                    [-1, 0, 0, 0, 0, 0, 0, 0, 0, 0]]),
+         jnp.array([4, 0, 0, 7]),
+         jnp.array([[4, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+                    [-1, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+                    [-1, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+                    [7, 0, 0, 0, 0, 0, 0, 0, 0, 0]])),
+    )
+    def test_vmap_update_node_resources(self, node_resource_array, selected_nodes, expected):
+        updated_node_resource_array = self.variant(vmap_update_node_resources)(node_resource_array, selected_nodes)
+        chex.assert_trees_all_close(updated_node_resource_array, expected)
 
 
 class ImplementNodeActionTest(parameterized.TestCase):
